gcf_rowAccessPolicies/requirements.txt
# Function dependencies, for example:
# package>=version
requests==2.28.2
google-auth==2.17.3
